SyntaxAnalyzer (in progress).
