+++ dbms/src/Interpreters/ArrayJoinedColumnsVisitor.h
@@ -1,5 +1,14 @@
 #pragma once
 
+#include <Core/Names.h>
+#include <Common/typeid_cast.h>
+
+#include <Parsers/ASTSubquery.h>
+#include <Parsers/ASTTablesInSelectQuery.h>
+#include <Parsers/ASTIdentifier.h>
+
+#include <DataTypes/NestedUtils.h>
+
 namespace DB
 {
