+++ dbms/src/Interpreters/ExpressionAnalyzer.h
@@ -3,6 +3,7 @@
 #include <Interpreters/AggregateDescription.h>
 #include <Interpreters/Settings.h>
 #include <Interpreters/ActionsVisitor.h>
+#include <Interpreters/SyntaxAnalyzer.h>
 
 namespace DB
 {
@@ -229,70 +230,7 @@ private:
     size_t subquery_depth;
     bool do_global; /// Do I need to prepare for execution global subqueries when analyzing the query.
 
-    struct AnalyzedJoin
-    {
-
-        /// NOTE: So far, only one JOIN per query is supported.
-
-        /** Query of the form `SELECT expr(x) AS k FROM t1 ANY LEFT JOIN (SELECT expr(x) AS k FROM t2) USING k`
-          * The join is made by column k.
-          * During the JOIN,
-          *  - in the "right" table, it will be available by alias `k`, since `Project` action for the subquery was executed.
-          *  - in the "left" table, it will be accessible by the name `expr(x)`, since `Project` action has not been executed yet.
-          * You must remember both of these options.
-          *
-          * Query of the form `SELECT ... from t1 ANY LEFT JOIN (SELECT ... from t2) ON expr(t1 columns) = expr(t2 columns)`
-          *     to the subquery will be added expression `expr(t2 columns)`.
-          * It's possible to use name `expr(t2 columns)`.
-          */
-        Names key_names_left;
-        Names key_names_right; /// Duplicating names are qualified.
-        ASTs key_asts_left;
-        ASTs key_asts_right;
-
-        struct JoinedColumn
-        {
-            /// Column will be joined to block.
-            NameAndTypePair name_and_type;
-            /// original column name from joined source.
-            String original_name;
-
-            JoinedColumn(const NameAndTypePair & name_and_type_, const String & original_name_)
-                    : name_and_type(name_and_type_), original_name(original_name_) {}
-
-            bool operator==(const JoinedColumn & o) const
-            {
-                return name_and_type == o.name_and_type && original_name == o.original_name;
-            }
-        };
-
-        using JoinedColumnsList = std::list<JoinedColumn>;
-
-        /// All columns which can be read from joined table. Duplicating names are qualified.
-        JoinedColumnsList columns_from_joined_table;
-        /// Columns which will be used in query to the joined query. Duplicating names are qualified.
-        NameSet required_columns_from_joined_table;
-
-        /// Columns which will be added to block, possible including some columns from right join key.
-        JoinedColumnsList columns_added_by_join;
-        /// Such columns will be copied from left join keys during join.
-        NameSet columns_added_by_join_from_right_keys;
-        /// Actions which need to be calculated on joined block.
-        ExpressionActionsPtr joined_block_actions;
-
-        void createJoinedBlockActions(const NamesAndTypesList & source_columns,
-                                      const ASTSelectQuery * select_query_with_join,
-                                      const Context & context);
-
-        NamesAndTypesList getColumnsAddedByJoin() const;
-
-        const JoinedColumnsList & getColumnsFromJoinedTable(const NamesAndTypesList & source_columns,
-                                                            const Context & context,
-                                                            const ASTSelectQuery * select_query_with_join);
-    };
-
     AnalyzedJoin analyzed_join;
-
 
     /** Remove all unnecessary columns from the list of all available columns of the table (`columns`).
       * At the same time, form a set of unknown columns (`unknown_required_source_columns`),
