+++ dbms/src/Interpreters/TranslateQualifiedNamesVisitor.h
@@ -5,24 +5,20 @@
 
 #include <Common/typeid_cast.h>
 #include <Parsers/DumpASTNode.h>
+#include <Parsers/ASTIdentifier.h>
+#include <Parsers/ASTQualifiedAsterisk.h>
+#include <Parsers/ASTSelectQuery.h>
+#include <Parsers/ASTTablesInSelectQuery.h>
 #include <Interpreters/DatabaseAndTableWithAlias.h>
 
 namespace DB
 {
 
-class ASTIdentifier;
-class ASTQualifiedAsterisk;
-class ASTSelectQuery;
-struct ASTTableJoin;
-
-class NamesAndTypesList;
-
-
 /// It visits nodes, find columns (general identifiers and asterisks) and translate their names according to tables' names.
 class TranslateQualifiedNamesVisitor
 {
 public:
-    TranslateQualifiedNamesVisitor(const NamesAndTypesList & source_columns_, const std::vector<DatabaseAndTableWithAlias> & tables_,
+    TranslateQualifiedNamesVisitor(const NameSet & source_columns_, const std::vector<DatabaseAndTableWithAlias> & tables_,
                                    std::ostream * ostr_ = nullptr)
     :   source_columns(source_columns_),
         tables(tables_),
@@ -42,7 +38,7 @@ public:
     }
 
 private:
-    const NamesAndTypesList & source_columns;
+    const NameSet & source_columns;
     const std::vector<DatabaseAndTableWithAlias> & tables;
     mutable size_t visit_depth;
     std::ostream * ostr;
